requirements.txt
asgiref==3.3.4
blinker==1.4
Brotli==1.0.9
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.3.2
cycler==0.10.0
decorator==4.4.2
Flask==1.1.4
flatbuffers==2.0
h11==0.12.0
h2==4.0.0
hpack==4.0.0
hyperframe==6.0.1
idna==2.10
imageio==2.9.0
itsdangerous==1.1.0
Jinja2==2.11.3
kaitaistruct==0.9
kiwisolver==1.3.1
ldap3==2.8.1
MarkupSafe==2.0.1
matplotlib==3.4.2
mitmproxy==6.0.2
msgpack==1.0.2
networkx==2.5.1
NudeNet==2.0.9
numpy==1.21.0
onnxruntime==1.8.0
opencv-python-headless==4.5.2.54
passlib==1.7.4
Pillow==8.2.0
progressbar2==3.53.1
protobuf==3.14.0
publicsuffix2==2.20191221
pyasn1==0.4.8
pycparser==2.20
pydivert==2.1.0
pydload==1.0.9
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyperclip==1.8.2
python-dateutil==2.8.1
python-utils==2.5.6
PyWavelets==1.1.1
requests==2.25.1
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.4
scikit-image==0.18.2
scipy==1.7.0
six==1.16.0
sortedcontainers==2.3.0
tifffile==2021.6.14
tornado==6.1
urllib3==1.26.6
urwid==2.1.2
Werkzeug==1.0.1
wsproto==1.0.0
zstandard==0.14.1
